create utils folder and refactor code
