+++ src/mock/task.js
@@ -1,5 +1,5 @@
 import {COLORS} from '../const.js';
-import {getRandomInteger} from '../utils.js';
+import {getRandomInteger} from '../utils/common.js';
 
 const generateDescription = () => {
   const DESCRIPTIONS = [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`];
