+++ src/utils/common.js
@@ -1,0 +1,6 @@
+export const getRandomInteger = (a = 0, b = 1) => {
+  const min = Math.ceil(Math.min(a, b));
+  const max = Math.floor(Math.max(a, b));
+
+  return Math.floor(min + Math.random() * (max - min + 1));
+};
